Move Route and Routes types in their own module
